Add Updated Row Count
- Add a row count to the database status view to display how many
rows have been added to the database since the last time the crawler
started. This adds some visual feedback to confirm that the database
is being updated properly. The row count performs a query every time
it is updated, thus insuring it is a real count of any row whose
timestamp is newer than the start time of the crawler.
- Add crawler-tools.js, a dispatcher that communicates between the
crawler object and other parts of the app
- Add format-date.js, a utility script which creates a consistent
date (String) format between the database and the app